requirements.txt
alpaca-py>=0.42.0
APScheduler>=3.10.4
exchange-calendars>=4.5.0
matplotlib>=3.8.0
numpy>=2.0.0
pandas>=2.2.0
python-dotenv>=1.0.1
requests>=2.31.0
SQLAlchemy>=2.0.0
tenacity>=9.0.0
yfinance>=0.2.40

